requirements.txt
face_recognition
pillow
numpy==1.24.3
#pip install -r .\requirements.txt